requirements.txt
numpy>=1.23
Pillow>=10.0
PyYAML>=6.0
matplotlib>=3.7
scipy>=1.10
pytest>=7.4
